app/requirements.txt
attrs==25.4.0
certifi==2026.1.4
charset-normalizer==3.4.4
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
cyclopts==4.5.4
docstring_parser==0.17.0
docutils==0.22.4
fonttools==4.61.1
idna==3.11
kiwisolver==1.4.9
markdown-it-py==4.0.0
matplotlib==3.10.8
mdurl==0.1.2
numpy==2.4.2
numpy-quaternion==2024.0.13
packaging==25.0
pillow==12.1.1
platformdirs==4.9.2
polars==1.38.1
polars-runtime-32==1.38.1
pooch==1.9.0
Pygments==2.19.2
pyparsing==3.3.2
PyQt6==6.10.2
PyQt6-Qt6==6.10.2
PyQt6_sip==13.11.0
pyqtgraph==0.14.0
python-dateutil==2.9.0.post0
pyvista==0.47.0
pyvistaqt==0.11.3
QtPy==2.4.3
requests==2.32.5
rich==14.3.3
rich-rst==1.3.2
scipy==1.17.0
scooby==0.11.0
setuptools==80.10.2
six==1.17.0
typing_extensions==4.15.0
urllib3==2.6.3
vtk==9.6.0
wheel==0.46.3
